requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-telegram-bot==20.3  # Вместо aiogram - работает везде
python-dotenv==1.0.0